Deprec cast_types and promote cast_Argo_variable_type
